providers: move `get_instance_name()` to providers.py
`get_instance_name()` is a snapcraft-specific call, so it is moved out
of the craft-providers interface.
