+++ snapcraft/providers/providers.py
@@ -1,0 +1,44 @@
+# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
+#
+# Copyright 2022 Canonical Ltd.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Snapcraft-specific code to interface with craft-providers."""
+
+from pathlib import Path
+
+
+def get_instance_name(
+    *, project_name: str, project_path: Path, build_on: str, build_for: str
+) -> str:
+    """Formulate the name for an instance using each of the given parameters.
+
+    Incorporate each of the parameters into the name to come up with a
+    predictable naming schema that avoids name collisions across multiple
+    projects.
+
+    :param project_name: Name of the project.
+    :param project_path: Directory of the project.
+    """
+    return "-".join(
+        [
+            "snapcraft",
+            project_name,
+            "on",
+            build_on,
+            "for",
+            build_for,
+            str(project_path.stat().st_ino),
+        ]
+    )
